update acceptance tests to use new signatures.  Create new acceptance… (#510)
* update acceptance tests to use new signatures.  Create new acceptance test to exercise old create token function signatures


* increase fees for account


* update description of HTS Precompile V1 test


* adjust fees


* increase fee for v1 acceptance tests


* increase fee for v1 acceptance tests once more


* address review comments


* refactor tests

+++ packages/server/tests/contracts/contracts_v1/IHederaTokenService.sol
@@ -140,7 +140,7 @@ interface IHederaTokenService {
         // IWA Compatibility. Depends on TokenSupplyType. For tokens of type FUNGIBLE_COMMON - the
         // maximum number of tokens that can be in circulation. For tokens of type NON_FUNGIBLE_UNIQUE -
         // the maximum number of NFTs (serial numbers) that can be minted. This field can never be changed!
-        int64 maxSupply;
+        uint32 maxSupply;
 
         // The default Freeze status (frozen or unfrozen) of Hedera accounts relative to this token. If
         // true, an account must be unfrozen before it can receive the token
@@ -359,8 +359,8 @@ interface IHederaTokenService {
     /// @return tokenAddress the created token's address
     function createFungibleToken(
         HederaToken memory token,
-        uint64 initialTotalSupply,
-        uint32 decimals)
+        uint initialTotalSupply,
+        uint decimals)
     external payable returns (int responseCode, address tokenAddress);
 
     /// Creates a Fungible Token with the specified properties
@@ -374,8 +374,8 @@ interface IHederaTokenService {
     /// @return tokenAddress the created token's address
     function createFungibleTokenWithCustomFees(
         HederaToken memory token,
-        uint64 initialTotalSupply,
-        uint32 decimals,
+        uint initialTotalSupply,
+        uint decimals,
         FixedFee[] memory fixedFees,
         FractionalFee[] memory fractionalFees)
     external payable returns (int responseCode, address tokenAddress);
